Fix issue #4: give SHA256SUM files a date suffix
As described in issue #4, the SHA256SUM files are not removed after creating an ISO. This is intentional so they do not have to be downloaded more than once. However, this means that they can cause checksum verification to fail if the same files are used to validate the ISO from a different date. I have added a date suffix (DD-MM-YYYY) to the SHA256SUMS and SHA256SUMS.gpg files so that they are only used for the current day.

There is still an issue if a user specifies an existing ISO image but does not disable automatic GPG verification (using the -k or --no-verify option), because if the specified image is not the current latest daily ISO, the verification will still fail. If users wish to use an existing ISO image other than the current latest one, they should disable GPG verification with the command line option and instead verify the ISO image themselves if desired.