Add option to use release instead of daily ISO
I ran into an issue with one of the daily releases where the ISO
wouldn't boot correctly.  I had no issues with the release ISO and don't
mind the time required for the few updates.

However, I still wanted to do all the automatic verification so
specifying the source ISO required a manual step or constantly updating
the SHA256SUM files.

These changes add some variables that allow the default to remain on the
daily build, but can be updated for the release build without breaking
the verification.